Extract dom manipulation functions to separate module
